[Renaming] Move RenamePropertyRector to appropriate set
