+++ packages/documentation-generator/tests/Printer/RectorsDocumentationPrinter/RectorsDocumentationPrinterTest.php
@@ -9,8 +9,8 @@ use Rector\Core\Contract\Rector\RectorInterface;
 use Rector\Core\Exception\ShouldNotHappenException;
 use Rector\Core\HttpKernel\RectorKernel;
 use Rector\DocumentationGenerator\Printer\RectorsDocumentationPrinter;
-use Rector\Generic\Rector\PropertyFetch\RenamePropertyRector;
 use Rector\Php74\Rector\Property\TypedPropertyRector;
+use Rector\Renaming\Rector\PropertyFetch\RenamePropertyRector;
 use ReflectionClass;
 use Symplify\PackageBuilder\Tests\AbstractKernelTestCase;
 use Symplify\SmartFileSystem\SmartFileInfo;
