--- rules/generic/tests/Rector/PropertyFetch/RenamePropertyRector/Source/ClassWithProperties.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\Generic\Tests\Rector\PropertyFetch\RenamePropertyRector\Source;
-
-final class ClassWithProperties
-{
-
-}
