+++ rules/renaming/tests/Rector/PropertyFetch/RenamePropertyRector/Source/ClassWithProperties.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Renaming\Tests\Rector\PropertyFetch\RenamePropertyRector\Source;
+
+final class ClassWithProperties
+{
+
+}
